[controller] Clean up push monitor classes
