--- services/venice-controller/src/main/java/com/linkedin/venice/pushmonitor/HelixCVBasedPushMonitor.java
@@ -1,8 +1,0 @@
-package com.linkedin.venice.pushmonitor;
-
-/**
- * An implementation of {@link AbstractPushMonitor} that listens to Helix
- * Customized View to determine the push status
- */
-public class HelixCVBasedPushMonitor {
-}
